requirements.txt
annotated-types==0.7.0
anyio==4.9.0
azure-communication-callautomation==1.4.0
azure-communication-chat==1.3.0
azure-communication-identity==1.5.0
azure-core==1.35.0
azure-identity==1.23.0
certifi==2025.7.9
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cryptography==45.0.5
fastapi==0.115.14
gunicorn==23.0.0
h11==0.16.0
httptools==0.6.4
idna==3.10
isodate==0.7.2
msal==1.32.3
msal-extensions==1.3.1
msrest==0.7.1
oauthlib==3.3.1
packaging==25.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
python-dotenv==1.1.1
PyYAML==6.0.2
requests==2.32.4
requests-oauthlib==2.0.0
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
watchfiles==1.1.0
websockets==15.0.1
google-generativeai==0.8.3
reportlab==4.2.5
